task-1/requirements.txt
selenium-wire
blinker==1.7.0
requests
python-dotenv